+++ contracts/OpenSTValueInterface.sol
@@ -54,7 +54,7 @@ contract OpenSTValueInterface {
         string _symbol,
         string _name,
         uint256 _conversionRate,
-        uint256 _conversionRateDecimalFactor,
+        uint8 _conversionRateDecimalFactor,
         uint256 _chainIdUtility,
         address _stakingAccount,
         bytes32 _checkUuid)
